+++ src/app/app-config/app-config.interface.ts
@@ -1,0 +1,11 @@
+
+interface PredefinedCategories {
+  deposit: string;
+  withdraw: string;
+}
+
+export interface AppConfig {
+
+  categories: PredefinedCategories
+
+}